+++ libportal/print.c
@@ -48,7 +48,7 @@ typedef struct {
   guint signal_id;
   GTask *task;
   char *request_path;
-  guint cancelled_id;
+  gulong cancelled_id;
 } PrintCall;
 
 static void
